requirements.txt
pandas
jupyter
pytest
Pillow
python-dotenv